# docs/swagger.yaml
definitions:
  query.BoothAddQuery:
    description: 摊位增加请求模型
    properties:
      address:
        description: 摊位地址
        type: string
      latitude:
        description: 维度
        type: number
      longitude:
        description: 经度
        type: number
      name:
        description: 摊位名称
        type: string
      status:
        description: 摊位状态
        type: integer
      type:
        description: 摊位类型
        type: integer
      userId:
        description: 用户id
        type: integer
    type: object
...
  query.EditBoothQuery:
    description: 编辑摊位请求模型
    properties:
      address:
        description: 摊位地址
        type: string
      bId:
        description: 摊位id
        type: integer
      latitude:
        description: 维度
        type: number
      longitude:
        description: 经度
        type: number
      name:
        description: 摊位名称
        type: string
      type:
        description: 摊位类型
        type: integer
    type: object
...
  utils.Result:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 返回数据
      message:
        description: 消息
        type: string
    type: object
info:
  contact: {}
paths:
  /booth/add:
    post:
      consumes:
      - application/json
      description: 添加一个摊位的信息
      parameters:
      - description: 摊位信息
        in: body
        name: booth
        required: true
        schema:
          $ref: '#/definitions/query.BoothAddQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 添加摊位
      tags:
      - 摊位模块
  /booth/detail:
    get:
      consumes:
      - application/json
      description: 获取一个摊位的详情
      parameters:
      - description: 摊位id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 摊位详情
      tags:
      - 摊位模块
  /booth/edit:
    post:
      consumes:
      - application/json
      description: 编辑一个摊位的信息
      parameters:
      - description: 摊位信息
        in: body
        name: booth
        required: true
        schema:
          $ref: '#/definitions/query.EditBoothQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 编辑摊位
      tags:
      - 摊位模块
  /booth/page:
    get:
      consumes:
      - application/json
      description: 获取摊位列表
      parameters:
      - description: 页号
        in: query
        name: pageNo
        type: integer
      - description: 页数
        in: query
        name: pageSize
        type: integer
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 类型
        in: query
        name: type
        type: integer
      - description: 摊主
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 摊位列表
      tags:
      - 摊位模块
...
  /file/add:
    post:
      consumes:
      - application/json
      description: 添加一个文件
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 添加文件
      tags:
      - 文件模块
  /file/del:
    post:
      consumes:
      - application/json
      description: 根据文件id删除文件
      parameters:
      - description: 文件id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除文件
      tags:
      - 文件模块
  /file/detail:
    get:
      consumes:
      - application/json
      description: 根据文件id获取文件详情
      parameters:
      - description: 文件id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 文件详情
      tags:
      - 文件模块
...
swagger: "2.0"